refactor: move common navigation behavior to ViewModelBase
