+++ Agendai/ViewModels/AgendaWindowViewModel.cs
@@ -5,16 +5,6 @@ namespace Agendai.ViewModels;
 
 public class AgendaWindowViewModel : ViewModelBase
 {
-    public MainWindowViewModel? MainViewModel { get; set; }
-
-    private ICommand? _returnHomeCommand;
-
-    public ICommand ReturnHomeCommand => _returnHomeCommand ??= new RelayCommand(ReturnHome);
-
-    private void ReturnHome()
-    {
-        MainViewModel?.NavigateToHome();
-    }
 
     public string Title { get; set; } = "Agenda";
 }
