+++ Agendai/ViewModels/HomeWindowViewModel.cs
@@ -5,8 +5,6 @@ namespace Agendai.ViewModels;
 
 public class HomeWindowViewModel : ViewModelBase
 {
-    public MainWindowViewModel? MainViewModel { get; set; }
-
     private ICommand? _openAgendaCommand;
 
     public ICommand OpenAgendaCommand => _openAgendaCommand ??= new RelayCommand(OpenAgenda);
