+++ Agendai/ViewModels/TodoWindowViewModel.cs
@@ -1,0 +1,6 @@
+namespace Agendai.ViewModels;
+
+public class TodoWindowViewModel : ViewModelBase
+{
+    public string Title { get; set; } = "Tarefas";
+}
