+++ Agendai/Views/Windows/TodoWindow/TodoWindow/TodoWindow.axaml
@@ -1,0 +1,12 @@
+<UserControl xmlns="https://github.com/avaloniaui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:vm="clr-namespace:Agendai.ViewModels"
+             x:Class="Agendai.TodoWindow"
+             x:DataType="vm:TodoWindowViewModel">
+  <StackPanel Spacing="10" Margin="20">
+        <TextBlock Text="{Binding Title}" FontSize="24"/>
+        <Button Command="{Binding ReturnHomeCommand}">Voltar para Home</Button>
+    </StackPanel>
+</UserControl>
