+++ Agendai/Views/Windows/TodoWindow/TodoWindow/TodoWindow.axaml.cs
@@ -1,0 +1,11 @@
+using Avalonia.Controls;
+
+namespace Agendai;
+
+public partial class TodoWindow : UserControl
+{
+    public TodoWindow()
+    {
+        InitializeComponent();
+    }
+}
